backend/pyproject.toml
[project]
name = "aischeduler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.11",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.39",
]
